geolocation/geo-watchPosition-015.html

<!DOCTYPE html>
<html>
  <head>
    <title>Geolocation Test: call watchPosition with options timeout 60000 and enableHighAccuracy false, then get Coordinates object</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#geolocation">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#watch-position">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script type="text/javascript">
      var t = async_test("Test checks that call watchPosition with options timeout 60000 and enableHighAccuracy false, then get Coordinates object");
      setup({timeout: 65000});
      try {
        navigator.geolocation.watchPosition(successCallback, errorCallback, { timeout : 60000, maximumAge : 0,  enableHighAccuracy : false});
      } catch (err) {
        t.step(function () {
          assert_unreached("Error message: " + err.message);
        });
        t.done();
      }
      function successCallback (position) {
        t.step(function () {
          assert_equals(typeof position.coords, "object", "call watchPosition with options timeout 60000 and enableHighAccuracy false, then get Coordinates object");
        });
        t.done();
      }
      function errorCallback (error) {
        t.step(function () {
          assert_unreached("Error message: " + error.message);
        });
        t.done();
      }
  </script>
</body>
</html>
